Rename schema to types
